+++ metrics/noop.go
@@ -1,4 +1,4 @@
-package metrics2
+package metrics
 
 type NoopMetricsReporter struct{}
